Organize file and folder structure as agreed
