+++ src/router.jsx
@@ -6,7 +6,7 @@ import MainLayout from './layouts/MainLayout';
 import PublicLayout from './layouts/PublicLayout';
 import MyProfile from './sections/myProfile/MyProfile';
 import LoginFormContainer from './sections/login/LoginFormContainer';
-import DestinationContainer from './sections/admin/destinations/containers/DestinationContainer';
+import Destination from './sections/admin/destinations/destination';
 import Destinations from './sections/admin/destinations/Destinations';
 import SignUpForm from './sections/signup/SignUpForm';
 import SignupTrip from './sections/signupTrip/SignupTrip';
@@ -28,7 +28,7 @@ export default(
                 <Route path="/profile" component={MyProfile} />
                 <Route path="/trips/signup" component={SignupTrip} />
                 <Route path="/admin/destinations" component={Destinations} />
-                <Route path="/admin/destinations/:destinationId" component={DestinationContainer} />
+                <Route path="/admin/destinations/:destinationId" component={Destination} />
                 <Route path="/admin/trips" component={TripRequests} />
                 <Route path="*" component={NotFound} />
             </Route>
